readwrite: move all extra type-info macros into one place
