+++ readWrite/src/main/scala-3/slinky/readwrite/CoreReadersMacro.scala
@@ -9,11 +9,11 @@ trait MacroReaders {
   inline implicit def deriveReader[T]: Reader[T] = {
     summonFrom {
       case r: Reader[T] => r
-      case vc: ValueClass[T] =>
+      case vc: ExoticTypes.ValueClass[T] =>
         MacroReaders.ValueClassReader(vc, summonInline[Reader[vc.Repr]])
       case m: Mirror.ProductOf[T] => deriveProduct(m)
       case m: Mirror.SumOf[T] => deriveSum(m)
-      case nu: NominalUnion[T] => MacroReaders.UnionReader(summonAll[Tuple.Map[nu.Constituents, Reader]])
+      case nu: ExoticTypes.NominalUnion[T] => MacroReaders.UnionReader(summonAll[Tuple.Map[nu.Constituents, Reader]])
     }
   }
 
@@ -21,7 +21,7 @@ trait MacroReaders {
     val labels = constValueTuple[m.MirroredElemLabels]
     val readers = summonAll[Tuple.Map[m.MirroredElemTypes, Reader]]
     val defaults = summonFrom {
-      case d: slinky.readwrite.DefaultConstructorParameters[T] => d.values
+      case d: ExoticTypes.DefaultConstructorParameters[T] => d.values
       case _ => null
     }
     MacroReaders.ProductReader(m, labels, readers, defaults)
@@ -34,7 +34,7 @@ trait MacroReaders {
 }
 
 object MacroReaders {
-  class ValueClassReader[T, R](vc: ValueClass[T] { type Repr = R }, reader: Reader[R]) extends Reader[T] {
+  class ValueClassReader[T, R](vc: ExoticTypes.ValueClass[T] { type Repr = R }, reader: Reader[R]) extends Reader[T] {
     protected def forceRead(o: scala.scalajs.js.Object): T = vc.to(reader.read(o))
   }
 
@@ -74,67 +74,3 @@ object MacroReaders {
     }
   }
 }
-//class MacroReadersImpl(_c: whitebox.Context) extends GenericDeriveImpl(_c) {
-//  import c.universe._
-//
-//  val typeclassType: c.universe.Type = typeOf[Reader[_]]
-//
-//  def deferredInstance(forType: c.universe.Type, constantType: c.universe.Type) =
-//    q"new _root_.slinky.readwrite.DeferredReader[$forType, $constantType]"
-//
-//  def maybeExtractDeferred(tree: c.Tree): Option[c.Tree] =
-//    tree match {
-//      case q"new _root_.slinky.readwrite.DeferredReader[$_, $t]()" =>
-//        Some(t)
-//      case q"new slinky.readwrite.DeferredReader[$_, $t]()" =>
-//        Some(t)
-//      case _ => None
-//    }
-//
-//  def createModuleTypeclass(tpe: c.universe.Type, moduleReference: c.Tree): c.Tree =
-//    q"""new _root_.slinky.readwrite.Reader[$tpe] {
-//          def forceRead(o: _root_.scala.scalajs.js.Object): $tpe = {
-//            $moduleReference
-//          }
-//        }"""
-//
-//  def createCaseClassTypeclass(clazz: c.Type, params: Seq[Seq[Param]]): c.Tree = {
-//    val paramsTrees = params.map(_.map { p =>
-//      p.transformIfVarArg {
-//        p.default.map { d =>
-//          q"if (_root_.scala.scalajs.js.isUndefined(o.asInstanceOf[_root_.scala.scalajs.js.Dynamic].${p.name.toTermName})) $d else ${getTypeclass(p.tpe)}.read(o.asInstanceOf[_root_.scala.scalajs.js.Dynamic].${p.name.toTermName}.asInstanceOf[_root_.scala.scalajs.js.Object])"
-//        }.getOrElse {
-//          q"${getTypeclass(p.tpe)}.read(o.asInstanceOf[_root_.scala.scalajs.js.Dynamic].${p.name.toTermName}.asInstanceOf[_root_.scala.scalajs.js.Object])"
-//        }
-//      }
-//    })
-//
-//    q"""new _root_.slinky.readwrite.Reader[$clazz] {
-//          def forceRead(o: _root_.scala.scalajs.js.Object): $clazz = {
-//            new $clazz(...$paramsTrees)
-//          }
-//        }"""
-//  }
-//
-//  def createValueClassTypeclass(clazz: c.Type, param: Param): c.Tree =
-//    q"""new _root_.slinky.readwrite.Reader[$clazz] {
-//          def forceRead(o: _root_.scala.scalajs.js.Object): $clazz = {
-//            new $clazz(${getTypeclass(param.tpe)}.read(o))
-//          }
-//        }"""
-//
-//  def createSealedTraitTypeclass(traitType: c.Type, subclasses: Seq[c.Symbol]): c.Tree = {
-//    val cases = subclasses.map(sub => cq"""${sub.name.toString} => ${getTypeclass(sub.asType.toType)}.read(o)""")
-//
-//    q"""new _root_.slinky.readwrite.Reader[$traitType] {
-//          def forceRead(o: _root_.scala.scalajs.js.Object): $traitType = {
-//            o.asInstanceOf[_root_.scala.scalajs.js.Dynamic]._type.asInstanceOf[_root_.java.lang.String] match {
-//              case ..$cases
-//              case _ => _root_.slinky.readwrite.Reader.fallback[$traitType].read(o)
-//            }
-//          }
-//        }"""
-//  }
-//
-//  def createFallback(forType: c.Type) = q"_root_.slinky.readwrite.Reader.fallback[$forType]"
-//}
